backend/requirements.txt
# 核心框架
fastapi>=0.104.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0

# 数据源
akshare>=1.0.0
tushare>=1.2.89

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 向量存储
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# LLM 接口
anthropic>=0.50.0
openai>=1.0.0

# 情绪分析 NLP（可选：首次使用时自动下载模型）
transformers>=4.35.0

# 深度学习预测（可选）
# torch>=2.0.0
# lightgbm>=4.0.0

# 用户认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
